chore: Merge `MemberController` into `MemberService`
