+++ crates/rest-api/src/error/errors.rs
@@ -4,7 +4,7 @@
 // License: Mozilla Public License v2.0 (MPL v2.0)
 
 use service::{
-    members::{member_service, UserCreateError, UserDeleteError},
+    members::{UserCreateError, UserDeleteError},
     notifications::notifier,
     sea_orm,
     xmpp::{server_ctl, server_manager, xmpp_service, CreateServiceAccountError},
@@ -200,16 +200,6 @@ impl CustomErrorCode for MutationError {
     }
 }
 impl_into_error!(MutationError);
-
-impl CustomErrorCode for member_service::Error {
-    fn error_code(&self) -> ErrorCode {
-        match self {
-            Self::CouldNotCreateUser(err) => err.code(),
-            Self::CouldNotDeleteUser(err) => err.code(),
-        }
-    }
-}
-impl_into_error!(member_service::Error);
 
 impl CustomErrorCode for UserCreateError {
     fn error_code(&self) -> ErrorCode {
