+++ crates/rest-api/src/features/members/get_member.rs
@@ -4,7 +4,7 @@
 // License: Mozilla Public License v2.0 (MPL v2.0)
 
 use rocket::{get, serde::json::Json};
-use service::{auth::UserInfo, members::MemberController};
+use service::{auth::UserInfo, members::MemberService};
 
 use crate::{
     error::{self, Error},
@@ -17,15 +17,15 @@ use super::EnrichedMember;
 #[get("/v1/members/<jid>")]
 pub async fn get_member_route<'r>(
     jid: JIDUriParam,
-    member_controller: LazyGuard<MemberController>,
+    member_service: LazyGuard<MemberService>,
     user_info: LazyGuard<UserInfo>,
 ) -> Result<Json<EnrichedMember>, Error> {
     // Make sure the user is logged in.
     let _ = user_info.inner?;
 
-    let member_controller = member_controller.inner?;
+    let member_service = member_service.inner?;
 
-    let member = member_controller.enrich_member(&jid).await?;
+    let member = member_service.enrich_member(&jid).await?;
     let Some(member) = member else {
         return Err(Error::from(error::NotFound {
             reason: format!("No member with id '{jid}'"),
