+++ crates/rest-api/src/features/members/get_members.rs
@@ -5,7 +5,7 @@
 
 use chrono::{DateTime, Utc};
 use rocket::get;
-use service::{auth::UserInfo, members::MemberController};
+use service::{auth::UserInfo, members::MemberService};
 
 use crate::{error::Error, forms::Timestamp, guards::LazyGuard, responders::Paginated};
 
@@ -13,7 +13,7 @@ use super::model::*;
 
 #[get("/v1/members?<page_number>&<page_size>&<until>")]
 pub async fn get_members_route<'r>(
-    member_controller: LazyGuard<MemberController>,
+    member_service: LazyGuard<MemberService>,
     user_info: LazyGuard<UserInfo>,
     page_number: Option<u64>,
     page_size: Option<u64>,
@@ -22,7 +22,7 @@ pub async fn get_members_route<'r>(
     // Make sure the user is logged in.
     let _ = user_info.inner?;
 
-    let member_controller = member_controller.inner?;
+    let member_service = member_service.inner?;
     let page_number = page_number.unwrap_or(1);
     let page_size = page_size.unwrap_or(20);
     let until: Option<DateTime<Utc>> = match until {
@@ -30,7 +30,7 @@ pub async fn get_members_route<'r>(
         None => None,
     };
 
-    let (pages_metadata, members) = member_controller
+    let (pages_metadata, members) = member_service
         .get_members(page_number, page_size, until)
         .await?;
