+++ crates/rest-api/src/features/members/guards/member_service.rs
@@ -5,20 +5,27 @@
 
 use std::sync::Arc;
 
-use service::{members::MemberController, xmpp::XmppService};
+use service::{
+    members::MemberService,
+    xmpp::{ServerCtl, XmppService},
+};
 
 use crate::guards::prelude::*;
 
 #[rocket::async_trait]
-impl<'r> LazyFromRequest<'r> for MemberController {
+impl<'r> LazyFromRequest<'r> for MemberService {
     type Error = error::Error;
 
     async fn from_request(req: &'r Request<'_>) -> Outcome<Self, Self::Error> {
         let db = try_outcome!(database_connection(req).await);
+        let server_ctl = try_outcome!(request_state!(req, ServerCtl));
         let xmpp_service = try_outcome!(XmppService::from_request(req).await);
 
-        Outcome::Success(MemberController::new(
+        try_outcome!(check_caller_is_admin(req, Some(db)).await);
+
+        Outcome::Success(MemberService::new(
             Arc::new(db.clone()),
+            Arc::new(server_ctl.clone()),
             Arc::new(xmpp_service),
         ))
     }
