+++ crates/rest-api/src/features/members/guards/mod.rs
@@ -3,4 +3,4 @@
 // Copyright: 2024, Rémi Bardon <remi@remibardon.name>
 // License: Mozilla Public License v2.0 (MPL v2.0)
 
-mod member_controller;
+mod member_service;
